+++ src/Plundi.Hammerfall.Core/Models/Abilities/Earthbreaker.cs
@@ -2,30 +2,30 @@ namespace Plundi.Hammerfall.Core.Models.Abilities;
 
 public class Earthbreaker : IAbility
 {
-    private readonly Dictionary<AbilityRarity, double> _cooldowns = new()
+    private readonly Dictionary<AbilityRarity, decimal> _cooldowns = new()
     {
-        { AbilityRarity.Common, 16 },
-        { AbilityRarity.Uncommon, 14 },
-        { AbilityRarity.Rare, 12 },
-        { AbilityRarity.Epic, 10 }
+        { AbilityRarity.Common, 16m },
+        { AbilityRarity.Uncommon, 14m },
+        { AbilityRarity.Rare, 12m },
+        { AbilityRarity.Epic, 10m }
     };
 
     private readonly Dictionary<AbilityRarity, DamageProfile> _damageProfiles = new()
     {
-        { AbilityRarity.Common, new() { BaseHits = [new() { Damage = 0.621, IsRelative = true, Timing = 2 }], SpecialHits = [], DotHits = [] } },
-        { AbilityRarity.Uncommon, new() { BaseHits = [new() { Damage = 0.663, IsRelative = true, Timing = 2 }], SpecialHits = [], DotHits = [] } },
-        { AbilityRarity.Rare, new() { BaseHits = [new() { Damage = 0.7, IsRelative = true, Timing = 2 }], SpecialHits = [], DotHits = [] } },
-        { AbilityRarity.Epic, new() { BaseHits = [new() { Damage = 0.73, IsRelative = true, Timing = 2 }], SpecialHits = [], DotHits = [] } }
+        { AbilityRarity.Common, new() { BaseHits = [new() { Damage = 0.621m, IsRelative = true, Timing = 2m }], SpecialHits = [], DotHits = [] } },
+        { AbilityRarity.Uncommon, new() { BaseHits = [new() { Damage = 0.663m, IsRelative = true, Timing = 2m }], SpecialHits = [], DotHits = [] } },
+        { AbilityRarity.Rare, new() { BaseHits = [new() { Damage = 0.7m, IsRelative = true, Timing = 2m }], SpecialHits = [], DotHits = [] } },
+        { AbilityRarity.Epic, new() { BaseHits = [new() { Damage = 0.73m, IsRelative = true, Timing = 2m }], SpecialHits = [], DotHits = [] } }
     };
 
     /// <inheritdoc />
     public string Name => "Earthbreaker";
 
     /// <inheritdoc />
-    public double CastDuration => 0;
+    public decimal CastDuration => 0m;
 
     /// <inheritdoc />
-    public double ChannelDuration => 2;
+    public decimal ChannelDuration => 2;
 
     /// <inheritdoc />
     public string ImagePath => "earthbreaker.jpg";
@@ -38,16 +38,16 @@ public class Earthbreaker : IAbility
     {
         return
         [
-            new() { Description = "AoE", Duration = 0 },
-            new() { Description = "immune to cc while channeling", Duration = 2 },
-            new() { Description = "stuns on hit", Duration = 2 },
-            new() { Description = "big knockback on hit", Duration = 0 },
-            new() { Description = "heavily slowed while channeling (-97%)", Duration = 2 }
+            new() { Description = "AoE", Duration = 0m },
+            new() { Description = "immune to cc while channeling", Duration = 2m },
+            new() { Description = "stuns on hit", Duration = 2m },
+            new() { Description = "big knockback on hit", Duration = 0m },
+            new() { Description = "heavily slowed while channeling (-97%)", Duration = 2m }
         ];
     }
 
     /// <inheritdoc />
-    public double GetCooldown(int characterLevel, AbilityRarity abilityRarity)
+    public decimal GetCooldown(int characterLevel, AbilityRarity abilityRarity)
     {
         return _cooldowns[abilityRarity];
     }
