+++ src/Plundi.Hammerfall.Core/Models/Abilities/Faeform.cs
@@ -2,12 +2,12 @@ namespace Plundi.Hammerfall.Core.Models.Abilities;
 
 public class Faeform : IAbility
 {
-    private readonly Dictionary<AbilityRarity, double> _cooldowns = new()
+    private readonly Dictionary<AbilityRarity, decimal> _cooldowns = new()
     {
-        { AbilityRarity.Common, 18 },
-        { AbilityRarity.Uncommon, 16 },
-        { AbilityRarity.Rare, 14 },
-        { AbilityRarity.Epic, 12 }
+        { AbilityRarity.Common, 18m },
+        { AbilityRarity.Uncommon, 16m },
+        { AbilityRarity.Rare, 14m },
+        { AbilityRarity.Epic, 12m }
     };
 
     private readonly Dictionary<AbilityRarity, DamageProfile> _damageProfiles = new()
@@ -22,10 +22,10 @@ public class Faeform : IAbility
     public string Name => "Faeform";
 
     /// <inheritdoc />
-    public double CastDuration => 0;
+    public decimal CastDuration => 0m;
 
     /// <inheritdoc />
-    public double ChannelDuration => 0;
+    public decimal ChannelDuration => 0;
 
     /// <inheritdoc />
     public string ImagePath => "faeform.jpg";
@@ -38,14 +38,14 @@ public class Faeform : IAbility
     {
         return
         [
-            new() { Description = "+125% movement speed", Duration = 3 },
-            new() { Description = "breaks all CC", Duration = 0 },
-            new() { Description = "immune to CC", Duration = 3 }
+            new() { Description = "+125% movement speed", Duration = 3m },
+            new() { Description = "breaks all CC", Duration = 0m },
+            new() { Description = "immune to CC", Duration = 3m }
         ];
     }
 
     /// <inheritdoc />
-    public double GetCooldown(int characterLevel, AbilityRarity abilityRarity)
+    public decimal GetCooldown(int characterLevel, AbilityRarity abilityRarity)
     {
         return _cooldowns[abilityRarity];
     }
