+++ src/Plundi.Hammerfall.Core/Models/AbilityEffect.cs
@@ -3,5 +3,5 @@ namespace Plundi.Hammerfall.Core.Models;
 public record AbilityEffect
 {
     public required string Description { get; init; }
-    public required double Duration { get; init; }
+    public required decimal Duration { get; init; }
 }
